Prepare fr movie nfo
