+++ src/main/java/kodi/nfo/Series.java
@@ -1,7 +1,6 @@
 package kodi.nfo;
 
 import kodi.common.UniqueId;
-import kodi.tvdb.TvDbArtworksResponse;
 import lombok.Builder;
 import lombok.Singular;
 import lombok.Value;
@@ -27,7 +26,6 @@ public class Series {
     @Builder.Default
     String tag = "anime";
 
-//    LocalDate added;
     LocalDate premiered;
     int year;
     String studio;
@@ -42,25 +40,4 @@ public class Series {
 
     @Singular
     List<Artwork> fanarts;
-
-    @Builder
-    @Value
-    public static class Artwork {
-        String url;
-        ArtworkType type;
-
-        @Builder.Default
-        int season = -1;
-    }
-
-    public enum ArtworkType {
-        SERIES_BANNER,
-        SERIES_POSTER,
-        SERIES_BACKGROUND,
-        SEASON_BANNER,
-        SEASON_POSTER,
-        SEASON_BACKGROUND,
-        CLEARART,
-        CLEARLOGO
-    }
 }
